Rename _dtype
