+++ geomstats/_backend/numpy/_dtype.py
