+++ geomstats/_backend/numpy/random.py
@@ -4,7 +4,7 @@ import numpy as _np
 from numpy.random import default_rng as _default_rng
 from numpy.random import randint, seed
 
-from ._dtype_wrapper import _add_default_dtype_by_casting
+from ._dtype import _add_default_dtype_by_casting
 
 rand = _add_default_dtype_by_casting(target=_np.random.rand)
 normal = _add_default_dtype_by_casting(target=_np.random.normal)
